% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adsmoothedladlasso.R
\name{adsmoothedladlasso}
\alias{adsmoothedladlasso}
\title{Multivariate adaptive smoothed LAD-lasso}
\usage{
adsmoothedladlasso(Y, X, tp, initialB = NULL, lambda = 0, h = 1, G = 10)
}
\arguments{
\item{Y}{an \eqn{n\times q} matrix of responses. The \eqn{i}th row contains the \eqn{q}-variate response
of the \eqn{i}th individual.}

\item{X}{an \eqn{n\times p} matrix of \eqn{p} explaining variables, The \eqn{i}th row contains the values
of \eqn{p} explaining variables for the \eqn{i}th individual.}

\item{tp}{a \eqn{q}-dimensional vector of time points.}

\item{initialB}{a \eqn{(p+1)\times q} matrix of initial regression coefficients}

\item{lambda}{the tuning parameter \eqn{\lambda} for the lasso penalty}

\item{h}{the bandwidth parameter}

\item{G}{the number of adaptive steps}
}
\value{
A list containing the following components:
\describe{
\item{beta}{the smoothed LAD-lasso regression coefficient matrix.}
\item{residuals}{the residuals.}
\item{lambda}{the tuning parameter \eqn{\lambda} for the LAD-penalty}
\item{iter}{the number of iterations}
\item{runtime}{the runtime of the function.}
}
}
\description{
\code{adsmoothedladlasso} is used to fit the multivariate fused LAD-lasso regression model.
}
\details{
Here are the details of the function...
}
\examples{
\dontrun{
data("simdat")
Y<-simdat[,1:2]
X<-simdat[,3:52]
out1<-adsmoothedladlasso(Y,X,lambda1=0,lambda2=0,G=5)
plot(out1)
out2<-adsmoothedladlasso(Y,X,lambda1=0.2,lambda2=0,G=5)
plot(out2)
out3<-adsmoothedladlasso(Y,X,lambda1=0.2,lambda2=0.2,G=5)
plot(out3)
}
}
\references{
Oja, H. (2010), \emph{Multivariate Nonparametric Methods with R.
An Approach Based on Spatial Signs and Ranks}, Springer.
\url{https://dx.doi.org/10.1007/978-1-4419-0468-3}.\cr
\cr
Nordhausen, K. and Oja, H. (2011), Multivariate L1 Methods:
The Package MNM, \emph{Journal of Statistical Software},
\strong{43}, 1-28. \url{https://doi.org/10.18637/jss.v043.i05}.
}
\seealso{
\code{\link{lambda1f.cv}} for cross-validation of lambda1.
}
