% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functionalladlasso.R
\name{functionalladlasso}
\alias{functionalladlasso}
\title{Multivariate functional LAD-lasso}
\usage{
functionalladlasso(
  Y,
  X,
  initialB = NULL,
  lambda1 = 0,
  lambda2 = 0,
  lpen = 1:dim(X)[2],
  method = "BFGS",
  gradient = FALSE,
  functional = 1,
  reltol = 1e-09,
  trace = 0
)
}
\arguments{
\item{Y}{an nxq matrix of responses. The ith row contains the q-variate response
of the ith individual.}

\item{X}{an nxp matrix of p explaining variables, The ith row contains the values
of p explaining variables for the ith individual.}

\item{initialB}{a (p+1)xq matrix of initial regression coefficients. If NULL, the initial values
are generated from standard normal distribution.}

\item{lambda1}{the tuning parameter \eqn{\lambda_1} for the lasso penalty.}

\item{lambda2}{the tuning parameter\eqn{\lambda_2} for the functional penalty.}

\item{lpen}{gives the lasso penalized coefficients. For example, lpen=c(2,5:8)
means that the coefficient vectors \eqn{\beta_2, \beta_5,...,\beta_8} are penalized.}

\item{method}{the optimization method to be used. The choices are "BFGS", "Nelder-Mead",
"CG", "L-BFGS-B", "SANN", "Brent" The default method is "BFGS". See the Details of the
function optim in package stats.}

\item{gradient}{a logical evaluating to TRUE or FALSE indicating whether gradient is used when method="BFGS".}

\item{functional}{functional penalty. If functional=1, then the functional penalty
is \eqn{\lambda_2\sum_{j=1}^p\sum_{k=2}^q|\beta_{j,k}-\beta_{j,k-1}|}. If functional=2,
then the functional penalty is \eqn{\lambda_2\sum_{k=2}^q||\beta^{(k)}-\beta^{(k-1)}||}.}

\item{reltol}{Relative convergence tolerance of the function optim.}

\item{trace}{Non-negative integer. If positive, tracing information on the progress of the optimization is produced.}
}
\value{
A list containing the following components:
\describe{
\item{beta}{the functional LAD regression coefficient matrix.}
\item{residuals}{the residuals.}
\item{lambda1}{the tuning parameter  \eqn{\lambda_1} for the LAD-lasso penalty}
\item{lambda2}{the tuning parameter  \eqn{\lambda_2} for the functional penalty}
\item{runtime}{the runtime of the function.}
\item{convergence}{convergence of the optimization routine. 0 indicates successful completion.}
\item{value}{the minimized value of the objective function}
}
}
\description{
\code{functionalladlasso} is used to fit the multivariate functional LAD-lasso regression model.
}
\details{
Here are the details of the function...
}
\examples{
\dontrun{
data("simdat")
Y<-simdat[,1:2]
X<-simdat[,3:32]
out<-functional.lad(Y,X,lambda1=0.2,lambda2=0.3)
out$runtime
}
}
\references{
Oja, H. (2010), \emph{Multivariate Nonparametric Methods with R.
An Approach Based on Spatial Signs and Ranks}, Springer.
\url{https://dx.doi.org/10.1007/978-1-4419-0468-3}.\cr
\cr
Nordhausen, K. and Oja, H. (2011), Multivariate L1 Methods:
The Package MNM, \emph{Journal of Statistical Software},
\strong{43}, 1-28. \url{https://doi.org/10.18637/jss.v043.i05}.
}
\seealso{
\code{\link{lambda1.cv}} for cross-validation of lambda1.
}
