% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l1.fit.R
\name{l1.fit}
\alias{l1.fit}
\title{L1 regression based on spatial signs}
\usage{
l1.fit(Y, X, initialB = NULL, maxiter = 1000, eps = 1e-06, eps.S = 1e-06)
}
\arguments{
\item{Y}{an nxq matrix of responses. The ith row contains the q-variate response
of the ith individual.}

\item{X}{an nx(p+1) design matrix.}

\item{initialB}{initial value of the coefficient matrix}

\item{maxiter}{maxiter}

\item{eps}{eps}

\item{eps.S}{eps.S}
}
\value{
A list containing the following components:
\describe{
\item{coefficients}{L1 estimate of the coefficient matrix}
\item{initial}{initial value of the coefficient matrix}
\item{iter}{the number of iterations}
\item{value}{the minimized value of the objective function}
}
}
\description{
\code{l1.fit} is used to find the multivariate L1 estimate for the coefficient matrix of the multivariate linear regression model
}
\details{
Here are the details of the function...
}
\references{
Oja, H. (2010), \emph{Multivariate Nonparametric Methods with R.
An Approach Based on Spatial Signs and Ranks}, Springer.
\url{https://dx.doi.org/10.1007/978-1-4419-0468-3}.\cr
\cr
Nordhausen, K. and Oja, H. (2011), Multivariate L1 Methods:
The Package MNM, \emph{Journal of Statistical Software},
\strong{43}, 1-28. \url{https://doi.org/10.18637/jss.v043.i05}.
}
\keyword{internal}
