% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lambda1s.cv.R
\name{lambda1s.cv}
\alias{lambda1s.cv}
\title{Selection of \eqn{\lambda} of the smoothed LAD-lasso using k-fold cross-validation}
\usage{
lambda1s.cv(
  Y,
  X,
  tp = 1:dim(Y)[2],
  lad = TRUE,
  lambda.min = 0,
  lambda.max = 5,
  len = 10,
  h = 1,
  k = 5
)
}
\arguments{
\item{Y}{an nxq matrix.}

\item{X}{an nxp matrix.}

\item{tp}{a \eqn{q}-dimensional vector of time points.}

\item{lad}{logical. If lad=TRUE, smoothed LAD-lasso is used, otherwise smoothed lasso is used.}

\item{lambda.min}{the minimum value of the grid of \eqn{\lambda}'s.}

\item{lambda.max}{the maximum value of the grid of \eqn{\lambda}'s.}

\item{len}{the number of values in the grid of \eqn{\lambda}'s}

\item{h}{the (fixed) value of the bandwidth parameter \eqn{h}.}

\item{k}{the number of folds.}
}
\value{
A list with the following components
\describe{
\item{lambda}{the grid of \eqn{\lambda}'s. The length of \eqn{lambda} is \eqn{len}}
\item{h}{the bandwidth parameter \eqn{h}.}
\item{cv}{list of values of CV precision measures in the grid points.
The first component vector gives the mean absolute errors (MAE) and the
second component vector gives the mean squared errors (MSE).}
\item{lbdmin}{vector of values of \eqn{\lambda} that minimize
the precision measures. The ith component
gives the value of \eqn{\lambda} that minimizes the ith
precision measure.}
\item{tpoint}{vector of types of values of \eqn{\lambda} that minimize
the precision measures: tpoint[i]=1, if the value of \eqn{\lambda} that
minimizes the ith precision measure is lambda1.min,
tpoint[i]=2, if the value of \eqn{\lambda} that minimizes the ith
precision measure is a turning point,  tpoint[i]=3, if the value
of \eqn{\lambda} that minimizes the ith precision measure is lambda1.max.}
\item{g}{the number of non-zero coefficient vectors.}
}
}
\description{
lambda1s.cv is used to find the tuning parameter  \eqn{\lambda} of the lasso penalty by using k-fold
cross-validation. The bandwidth parameter \eqn{h}  is fixed.
}
\details{
Here are the details of the function...
}
\examples{
\dontrun{
data("simdat")
Y<-simdat[,1:2]
X<-simdat[,3:52]
out <-lambda1s.cv(Y,X,lambda.min = 0.0001,lambda.max = 0.1,len=10,h=1,k=5)
plot(out)
}
}
\references{
Oja, H. (2010), \emph{Multivariate Nonparametric Methods with R.
An Approach Based on Spatial Signs and Ranks}, Springer.
\url{https://dx.doi.org/10.1007/978-1-4419-0468-3}.\cr
\cr
Nordhausen, K. and Oja, H. (2011), Multivariate L1 Methods:
The Package MNM, \emph{Journal of Statistical Software},
\strong{43}, 1-28. \url{https://doi.org/10.18637/jss.v043.i05}.
}
\seealso{
\code{\link{fusedladlasso}} for multivariate fused LAD-lasso.
}
