% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mladlasso-package.R
\docType{package}
\name{mladlasso-package}
\alias{mladlasso}
\alias{mladlasso-package}
\title{mladlasso: Multivariate LAD-lasso regression methods}
\description{
Multivariate LAD-lasso regression methods.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jymottonen/mladlasso}
  \item Report bugs at \url{https://github.com/jymottonen/mladlasso/issues}
}

}
\author{
\strong{Maintainer}: Jyrki Mottonen \email{jyrki.mottonen@helsinki.fi} (\href{https://orcid.org/0000-0002-6270-2556}{ORCID})

Authors:
\itemize{
  \item Jyrki Mottonen \email{jyrki.mottonen@helsinki.fi} (\href{https://orcid.org/0000-0002-6270-2556}{ORCID})
  \item Tero Lähderanta \email{Tero.Lahderanta@oulu.fi} (\href{https://orcid.org/0000-0001-5856-2899}{ORCID})
}

}
\keyword{internal}
