% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot.adfusedladlasso}
\alias{plot.adfusedladlasso}
\title{plot.adfusedladlasso}
\usage{
\method{plot}{adfusedladlasso}(
  x,
  p = x$p,
  q = x$q,
  k = 1,
  xlab = "Explaining variable",
  output = "screen",
  lambda1 = x$lambda1,
  lambda2 = x$lambda2,
  file = "Rplots.pdf",
  width = 7,
  height = 7,
  ...
)
}
\arguments{
\item{x}{an object of class fusedladlasso.}

\item{p}{number of explaining variables}

\item{q}{number of response variables}

\item{k}{plot of the kth iteration}

\item{xlab}{a title for the x axis. The default is "Explaining variable".}

\item{output}{the plot is exported to a PDF file \code{file} if \code{output}="PDF". Defaults \code{output}="screen".}

\item{lambda1}{the tuning parameter \eqn{\lambda_1} for the lasso penalty.}

\item{lambda2}{the tuning parameter \eqn{\lambda_2} for the fusion penalty.}

\item{file}{the name of the pdf file. The default is "Rplots.pdf".}

\item{width}{the width of the graphics region of the pdf file in inches. The default value is 7.}

\item{height}{the height of the graphics region of the pdf file in inches. The default value is 7.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Plot method for objects of class "adfusedladlasso".
}
\details{
Here are the details of the function...
}
