% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot.bic}
\alias{plot.bic}
\title{plot.bic}
\usage{
\method{plot}{bic}(x, ...)
}
\arguments{
\item{x}{an object of class bic.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Plot method for objects of class "bic".
}
\details{
Here are the details of the function...
}
