% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot.cv}
\alias{plot.cv}
\title{plot.cv}
\usage{
\method{plot}{cv}(x, cv = "MAE", ...)
}
\arguments{
\item{x}{an object of class cv.}

\item{cv}{the type of the plot. The choice \code{cv}="MAE" gives the cross-validation mean absolute errors (MAE)
and \code{cv}="MSE" gives the cross-validation mean squared errors (MSE).  The default is \code{cv}="MAE".}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Plot method for objects of class "cv".
}
\details{
Here are the details of the function...
}
