\name{simdat}
\alias{simdat}
\docType{data}
\title{
simdat
}
\description{
Simulated dataset
}
\usage{data("simdat")}
\format{
  A data frame with 100 observations on the following 32 variables.
  \describe{
    \item{\code{Y.1}}{a numeric vector}
    \item{\code{Y.2}}{a numeric vector}
    \item{\code{X.1}}{a numeric vector}
    \item{\code{X.2}}{a numeric vector}
    \item{\code{X.3}}{a numeric vector}
    \item{\code{X.4}}{a numeric vector}
    \item{\code{X.5}}{a numeric vector}
    \item{\code{X.6}}{a numeric vector}
    \item{\code{X.7}}{a numeric vector}
    \item{\code{X.8}}{a numeric vector}
    \item{\code{X.9}}{a numeric vector}
    \item{\code{X.10}}{a numeric vector}
    \item{\code{X.11}}{a numeric vector}
    \item{\code{X.12}}{a numeric vector}
    \item{\code{X.13}}{a numeric vector}
    \item{\code{X.14}}{a numeric vector}
    \item{\code{X.15}}{a numeric vector}
    \item{\code{X.16}}{a numeric vector}
    \item{\code{X.17}}{a numeric vector}
    \item{\code{X.18}}{a numeric vector}
    \item{\code{X.19}}{a numeric vector}
    \item{\code{X.20}}{a numeric vector}
    \item{\code{X.21}}{a numeric vector}
    \item{\code{X.22}}{a numeric vector}
    \item{\code{X.23}}{a numeric vector}
    \item{\code{X.24}}{a numeric vector}
    \item{\code{X.25}}{a numeric vector}
    \item{\code{X.26}}{a numeric vector}
    \item{\code{X.27}}{a numeric vector}
    \item{\code{X.28}}{a numeric vector}
    \item{\code{X.29}}{a numeric vector}
    \item{\code{X.30}}{a numeric vector}
  }
}
\keyword{datasets}
