% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stability.R
\name{stability}
\alias{stability}
\title{Stability selection}
\usage{
stability(
  Y,
  X,
  initialB = NULL,
  lambda1 = 0,
  lambda2 = 0,
  method = "BFGS",
  gradient = FALSE,
  functional = 0,
  reltol = 1e-08,
  trace = 0,
  N = 100,
  efp = 1
)
}
\arguments{
\item{Y}{an \eqn{n\times q} matrix of responses. The \eqn{i}th row contains the \eqn{q}-variate response
of the \eqn{i}th individual.}

\item{X}{an \eqn{n\times p} matrix of \eqn{p} explaining variables, The \eqn{i}th row contains the values
of \eqn{p} explaining variables for the \eqn{i}th individual.}

\item{initialB}{a \eqn{(p+1)\times q} matrix of initial regression coefficients. The default value is NULL.}

\item{lambda1}{the tuning parameter \eqn{\lambda_1} for the lasso penalty}

\item{lambda2}{the tuning parameter \eqn{\lambda_2} for the fusion penalty}

\item{method}{the optimization method to be used when functional=1 or 2 and \eqn{\lambda_2>0}. The choices are "BFGS", "Nelder-Mead",
"CG", "L-BFGS-B", "SANN", "Brent" The default method is "BFGS". See the Details of the
function optim in package stats.}

\item{gradient}{a logical evaluating to TRUE or FALSE indicating whether gradient is used when method="BFGS".}

\item{functional}{functional penalty. If functional=0, then the functional penalty is
\eqn{\lambda_2\sum_{j=2}^{p}||\beta_{j}-\beta_{j-1}||}.
If functional=1, then the functional penalty
is \eqn{\lambda_2\sum_{j=1}^p\sum_{k=2}^q|\beta_{j,k}-\beta_{j,k-1}|}. If functional=2,
then the functional penalty is \eqn{\lambda_2\sum_{k=2}^q||\beta^{(k)}-\beta^{(k-1)}||}.}

\item{reltol}{Relative convergence tolerance of the function optim.}

\item{trace}{Non-negative integer. If positive, tracing information on the progress of the optimization is produced.}

\item{N}{the number of replicates. The default value is 100.}

\item{efp}{the expected number of false positives we want to control. The default value is 1.}
}
\value{
A list containing the following components:
\describe{
\item{freq}{a q-vector,recording the stability frequencies of each marker over B replicates.}
\item{markersig}{the indices of markers which are judged to be significant by stability selection.}
}
}
\description{
\code{stability} is used to calculate the stability selection frequencies.
}
\details{
Here are the details of the function...
}
\references{
Oja, H. (2010), \emph{Multivariate Nonparametric Methods with R.
An Approach Based on Spatial Signs and Ranks}, Springer.
\url{https://dx.doi.org/10.1007/978-1-4419-0468-3}.\cr
\cr
Nordhausen, K. and Oja, H. (2011), Multivariate L1 Methods:
The Package MNM, \emph{Journal of Statistical Software},
\strong{43}, 1-28. \url{https://doi.org/10.18637/jss.v043.i05}.\cr
\cr
Li, Z., Mottonen, J. and Sillanpaa, M. J. (2015), A robust multiple-locus
method for quantitative trait locus analysis of non-normally distributed
multiple traits, \emph{Heredity}, \strong{115}, 556-564. \url{https://doi.org/10.1038/hdy.2015.61}.
}
