% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaries.R
\name{summary.fusedladlasso}
\alias{summary.fusedladlasso}
\title{summary.fusedladlasso}
\usage{
\method{summary}{fusedladlasso}(object, ..., digits = 3)
}
\arguments{
\item{object}{an object of class fusedladlasso.}

\item{...}{further arguments passed to or from other methods.}

\item{digits}{minimal number of significant digits.}
}
\description{
summary.fusedladlasso is used to print the summary of
themultivariate fused LAD-lasso regression fit.
}
\details{
Here are the details of the function...
}
