% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootladlasso.R
\name{bootladlasso}
\alias{bootladlasso}
\title{Bootstrapping adaptive fused LAD-lasso}
\usage{
bootladlasso(Y, X, initialB = NULL, lambda1 = 0, lambda2 = 0, K = 10)
}
\arguments{
\item{Y}{an \eqn{n\times q} matrix of responses. The \eqn{i}th row contains the \eqn{q}-variate response
of the \eqn{i}th individual.}

\item{X}{an \eqn{n\times p} matrix of \eqn{p} explaining variables, The \eqn{i}th row contains the values
of \eqn{p} explaining variables for the \eqn{i}th individual.}

\item{initialB}{a \eqn{(p+1)\times q} matrix of initial regression coefficients}

\item{lambda1}{the tuning parameter \eqn{\lambda_1} for the lasso penalty}

\item{lambda2}{the tuning parameter \eqn{\lambda_2} for the fusion penalty}

\item{K}{the number of adaptive steps}
}
\value{
A list containing the following components:
\describe{
\item{beta}{the fused LAD-lasso regression coefficient matrix.}
\item{residuals}{the residuals.}
\item{lambda1}{the tuning parameter \eqn{\lambda_1} for the LAD-penalty}
\item{lambda2}{the tuning parameter \eqn{\lambda_2} for the fusion-penalty}
\item{iter}{the number of iterations}
\item{runtime}{the runtime of the function.}
\item{convergence}{convergence of the optimization routine. 0 indicates successful completion.}
\item{value}{the minimized value of the objective function.}
}
}
\description{
\code{bootladlasso} is used to
}
\details{
Here are the details of the function...
}
\examples{
\dontrun{
data("simdat")
Y<-simdat[,1:2]
X<-simdat[,3:52]
out1<-adfusedladlasso(Y,X,lambda1=0,lambda2=0)
plot(out1)
out2<-adfusedladlasso(Y,X,lambda1=0.2,lambda2=0)
plot(out2)
out3<-adfusedladlasso(Y,X,lambda1=0.2,lambda2=0.2)
plot(out3)
}
}
\references{
Oja, H. (2010), \emph{Multivariate Nonparametric Methods with R.
An Approach Based on Spatial Signs and Ranks}, Springer.
\url{https://dx.doi.org/10.1007/978-1-4419-0468-3}.\cr
\cr
Nordhausen, K. and Oja, H. (2011), Multivariate L1 Methods:
The Package MNM, \emph{Journal of Statistical Software},
\strong{43}, 1-28. \url{https://doi.org/10.18637/jss.v043.i05}.
}
\seealso{
\code{\link{lambda1f.cv}} for cross-validation of lambda1.
}
