% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fusedladlasso.R
\name{fusedladlasso}
\alias{fusedladlasso}
\title{Multivariate fused LAD-lasso}
\usage{
fusedladlasso(Y, X, lambda1 = 0, lambda2 = 0)
}
\arguments{
\item{Y}{an nxq matrix}

\item{X}{an nxp matrix}

\item{lambda1}{the tuning parameter for the LAD-penalty}

\item{lambda2}{the tuning parameter for the fusion-penalty}
}
\value{
A list containing the following components:
\describe{
\item{beta}{the fused LAD-lasso regression coefficient matrix.}
\item{runtime}{the runtime of the function.}
}
}
\description{
\code{fusedladlasso} is used to fit the multivariate fused LAD-lasso regression model.
}
\details{
Here are the details of the function...
}
\examples{
\dontrun{
data("simdat")
Y<-simdat[,1:2]
X<-simdat[,3:32]
out<-fusedladlasso(Y,X,lambda1=0.2,lambda2=0)
out$runtime
}
}
\references{
Oja, H. (2010), \emph{Multivariate Nonparametric Methods with R.
An Approach Based on Spatial Signs and Ranks}, Springer.
\url{https://dx.doi.org/10.1007/978-1-4419-0468-3}.\cr
\cr
Nordhausen, K. and Oja, H. (2011), Multivariate L1 Methods:
The Package MNM, \emph{Journal of Statistical Software},
\strong{43}, 1-28. \url{https://doi.org/10.18637/jss.v043.i05}.
}
\seealso{
\code{\link{lambda1.cv}} for cross-validation of lambda1.
}
