% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fusedladlasso.R
\name{fusedladlasso}
\alias{fusedladlasso}
\title{Multivariate fused LAD-lasso}
\usage{
fusedladlasso(
  Y,
  X,
  initialB = NULL,
  lambda1 = 0,
  lambda2 = 0,
  method = "BFGS",
  gradient = FALSE,
  functional = 0,
  reltol = 1e-09,
  trace = 0
)
}
\arguments{
\item{Y}{an \eqn{n\times q} matrix of responses. The \eqn{i}th row contains the \eqn{q}-variate response
of the \eqn{i}th individual.}

\item{X}{an \eqn{n\times p} matrix of \eqn{p} explaining variables, The \eqn{i}th row contains the values
of \eqn{p} explaining variables for the \eqn{i}th individual.}

\item{initialB}{a \eqn{(p+1)\times q} matrix of initial regression coefficients}

\item{lambda1}{the tuning parameter \eqn{\lambda_1} for the lasso penalty}

\item{lambda2}{the tuning parameter \eqn{\lambda_2} for the fusion penalty}

\item{method}{the optimization method to be used when functional=1 or 2 and \eqn{\lambda_2>0}. The choices are "BFGS", "Nelder-Mead",
"CG", "L-BFGS-B", "SANN", "Brent" The default method is "BFGS". See the Details of the
function optim in package stats.}

\item{gradient}{a logical evaluating to TRUE or FALSE indicating whether gradient is used when method="BFGS".}

\item{functional}{functional penalty. If functional=0, then the functional penalty is
\eqn{\lambda_2\sum_{j=2}^{p}||\beta_{j}-\beta_{j-1}||}.
If functional=1, then the functional penalty
is \eqn{\lambda_2\sum_{j=1}^p\sum_{k=2}^q|\beta_{j,k}-\beta_{j,k-1}|}. If functional=2,
then the functional penalty is \eqn{\lambda_2\sum_{k=2}^q||\beta^{(k)}-\beta^{(k-1)}||}.}

\item{reltol}{Relative convergence tolerance of the function optim.}

\item{trace}{Non-negative integer. If positive, tracing information on the progress of the optimization is produced.}
}
\value{
A list containing the following components:
\describe{
\item{beta}{the fused LAD-lasso regression coefficient matrix.}
\item{residuals}{the residuals.}
\item{lambda1}{the tuning parameter \eqn{\lambda_1} for the LAD-penalty}
\item{lambda2}{the tuning parameter \eqn{\lambda_2} for the fusion-penalty}
\item{iter}{the number of iterations}
\item{runtime}{the runtime of the function.}
\item{convergence}{convergence of the optimization routine. 0 indicates successful completion.}
\item{value}{the minimized value of the objective function.}
}
}
\description{
\code{fusedladlasso} is used to fit the multivariate fused LAD-lasso regression model.
}
\details{
Here are the details of the function...
}
\examples{
\dontrun{
data("simdat")
Y<-simdat[,1:2]
X<-simdat[,3:52]
out1<-fusedladlasso(Y,X,lambda1=0,lambda2=0)
plot(out1)
out2<-fusedladlasso(Y,X,lambda1=0.2,lambda2=0)
plot(out2)
out3<-fusedladlasso(Y,X,lambda1=0.2,lambda2=0.2)
plot(out3)
out <-lambda1.cv(Y,X,lambda1.min = 0.0001,lambda1.max = 0.1,len1=10,lambda2 = 0)
out
}
}
\references{
Oja, H. (2010), \emph{Multivariate Nonparametric Methods with R.
An Approach Based on Spatial Signs and Ranks}, Springer.
\url{https://dx.doi.org/10.1007/978-1-4419-0468-3}.\cr
\cr
Nordhausen, K. and Oja, H. (2011), Multivariate L1 Methods:
The Package MNM, \emph{Journal of Statistical Software},
\strong{43}, 1-28. \url{https://doi.org/10.18637/jss.v043.i05}.
}
\seealso{
\code{\link{lambda1.cv}} for cross-validation of lambda1.
}
