% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fusedlasso.R
\name{fusedlasso}
\alias{fusedlasso}
\title{Multivariate fused lasso}
\usage{
fusedlasso(
  Y,
  X,
  lambda1 = 0,
  lambda2 = 0,
  lpen = 1:dim(X)[2],
  fpen = list(1:dim(X)[2])
)
}
\arguments{
\item{Y}{an nxq matrix of responses. The ith row contains the q-variate response
of the ith individual.}

\item{X}{an nxp matrix of p explaining variables, The ith row contains the values
of p explaining variables for the ith individual.}

\item{lambda1}{the tuning parameter for the lasso penalty}

\item{lambda2}{the tuning parameter for the fusion penalty}

\item{lpen}{gives the lasso penalized coefficients. For example, lpen=c(2,5:8)
means that the coefficient vectors \eqn{\beta_2, \beta_5,...,\beta_8} are penalized.}

\item{fpen}{a list of blocks of fusion penalized coefficients. For example,
fpen=list(2:5,10:20) means that the fusion penalty is
\eqn{\lambda_2[||\beta_3-\beta_2||+...+||\beta_5-\beta_4||+||\beta_{11}-\beta_{10}||+...+||\beta_{20}-\beta_{19}||}}
}
\value{
A list containing the following components:
\describe{
\item{beta}{the fused lasso regression coefficient matrix.}
\item{runtime}{the runtime of the function.}
}
}
\description{
\code{fusedlasso} is used to fit the multivariate fused lasso regression model.
}
\details{
Here are the details of the function...
}
\examples{
\dontrun{
data("simdat")
Y<-simdat[,1:2]
X<-simdat[,3:52]
out1<-fusedlasso(Y,X,lambda1=0,lambda2=0)
plot(out1)
out2<-fusedlasso(Y,X,lambda1=0.2,lambda2=0)
plot(out2)
out3<-fusedlasso(Y,X,lambda1=0.2,lambda2=0.2)
plot(out3)
}
}
\references{
Oja, H. (2010), \emph{Multivariate Nonparametric Methods with R.
An Approach Based on Spatial Signs and Ranks}, Springer.
\url{https://dx.doi.org/10.1007/978-1-4419-0468-3}.
}
\seealso{
\code{\link{lambda1.cv}} for cross-validation of lambda1.
}
