% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lambda1.cv.R
\name{lambda1.cv}
\alias{lambda1.cv}
\title{Cross-validation of lambda1}
\usage{
lambda1.cv(
  Y,
  X,
  lambda1.min = 0,
  lambda1.max = 5,
  len1 = 10,
  lambda2 = 0,
  lpen = 1:dim(X)[2],
  fpen = list(1:dim(X)[2])
)
}
\arguments{
\item{Y}{an nxq matrix.}

\item{X}{an nxp matrix.}

\item{lambda1.min}{the minimum value of the grid of lambda1's.}

\item{lambda1.max}{the maximum value of the grid of lambda1's.}

\item{len1}{the number of values in the grid of lambda1's}

\item{lambda2}{the (fixed) value of the tuning parameter lambda2.}

\item{lpen}{gives the lasso penalized coefficients. For example, lpen=c(2,5:8)
means that the coefficient vectors beta2, beta5,...,beta8 are penalized.}

\item{fpen}{a list of blocks of fusion penalized coefficients. For example,
fpen=list(2:5,10:20) means that the fusion penalty is
\code{lambda2*[||beta_3-beta_2||+...+||beta_5-beta_4||+
||beta_{11}-beta_{10}||+...+||beta_{20}-beta_{19}||]}}
}
\value{
A list with the following components
\describe{
\item{lambda1}{the grid of lambda1's. The length lambda1 is len1}
\item{lambda2}{the tuning parameter lambda2.}
\item{cv}{list of values of CV precision measures.
The ith component vector gives the values of the ith
precision measure in the grid points.}
\item{lbdmin}{vector of values of lambda1 that minimize
the precision measures. The ith component
gives the value of lambda1 that minimizes the ith
precision measure.}
\item{tpoint}{vector of types of values of lambda1 that minimize
the precision measures: tpoint[i]=1, if the value of lambda1 that
minimizes the ith precision measure is lambda1.min,
tpoint[i]=2, if the value of lambda1 that minimizes the ith
precision measure is a turning point,  tpoint[i]=3, if the value
of lambda1 that minimizes the ith precision measure is lambda1.max.}
\item{h}{the number non-zero coefficients.}
}
}
\description{
lambda1.cv is used to find the tuning parameter of the lasso penalty lambda1 by using 5-fold
cross-validation. The tuning parameter of the fusion penalty lambda2 is fixed.
}
\details{
Here are the details of the function...
}
\examples{
\dontrun{
data("simdat")
Y<-simdat[,1:2]
X<-simdat[,3:32]
out <-lambda1.cv(Y,X,lambda1.min = 0.0001,lambda1.max = 0.1,len1=10,lambda2 = 0)
plot(out)
}
}
\references{
Oja, H. (2010), \emph{Multivariate Nonparametric Methods with R.
An Approach Based on Spatial Signs and Ranks}, Springer.
\url{https://dx.doi.org/10.1007/978-1-4419-0468-3}.\cr
\cr
Nordhausen, K. and Oja, H. (2011), Multivariate L1 Methods:
The Package MNM, \emph{Journal of Statistical Software},
\strong{43}, 1-28. \url{https://doi.org/10.18637/jss.v043.i05}.
}
\seealso{
\code{\link{fusedladlasso}} for multivariate fused LAD-lasso.
}
