% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mladlasso-package.R
\docType{package}
\name{mladlasso-package}
\alias{mladlasso}
\alias{mladlasso-package}
\title{mladlasso: Multivariate LAD-lasso estimation}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Jyrki Mottonen \email{jyrki.mottonen@helsinki.fi} (\href{https://orcid.org/0000-0002-6270-2556}{ORCID})

}
\keyword{internal}
