% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot.cv}
\alias{plot.cv}
\title{plot.cv}
\usage{
\method{plot}{cv}(x, cv = 0, line = NA, ...)
}
\arguments{
\item{x}{an object of class fusedladlasso.}

\item{cv}{cv}

\item{line}{line}

\item{...}{further arguments passed to or from other methods.}
}
\description{
plot.cv is used to plot the results of
the cross-validation
}
\details{
Here are the details of the function...
}
