% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot.fusedladlasso}
\alias{plot.fusedladlasso}
\title{plot.fusedladlasso}
\usage{
\method{plot}{fusedladlasso}(x, ...)
}
\arguments{
\item{x}{an object of class fusedladlasso.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
plot.fusedladlasso is used to plot the results of
the  multivariate fused LAD-lasso regression fit.
}
\details{
Here are the details of the function...
}
