% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoothedladlasso.R
\name{smoothedladlasso}
\alias{smoothedladlasso}
\title{Multivariate smoothed LAD-lasso}
\usage{
smoothedladlasso(Y, X, tp = 1:dim(Y)[2], initialB = NULL, lambda = 0, h = 1)
}
\arguments{
\item{Y}{an \eqn{n\times q} matrix of responses. The \eqn{i}th row contains the \eqn{q}-variate response
of the \eqn{i}th individual.}

\item{X}{an \eqn{n\times p} matrix of \eqn{p} explaining variables, The \eqn{i}th row contains the values
of \eqn{p} explaining variables for the \eqn{i}th individual.}

\item{tp}{a \eqn{q}-dimensional vector of time points.}

\item{initialB}{a \eqn{(p+1)\times q} matrix of initial regression coefficients}

\item{lambda}{the tuning parameter \eqn{\lambda} for the lasso penalty}

\item{h}{the bandwidth parameter}
}
\value{
A list containing the following components:
\describe{
\item{beta}{the smoothed LAD-lasso regression coefficient matrix.}
\item{residuals}{the residuals.}
\item{lambda}{the tuning parameter \eqn{\lambda} for the LAD-penalty}
\item{iter}{the number of iterations}
\item{runtime}{the runtime of the function.}
}
}
\description{
\code{smoothedladlasso} is used to fit the multivariate fused LAD-lasso regression model.
}
\details{
Here are the details of the function...
}
\examples{
\dontrun{
data("simdat")
Y<-simdat[,1:2]
X<-simdat[,3:52]
out1<-smoothedladlasso(Y,X,lambda1=0,lambda2=0)
plot(out1)
out2<-smoothedladlasso(Y,X,lambda1=0.2,lambda2=0)
plot(out2)
out3<-smoothedladlasso(Y,X,lambda1=0.2,lambda2=0.2)
plot(out3)
}
}
\references{
Oja, H. (2010), \emph{Multivariate Nonparametric Methods with R.
An Approach Based on Spatial Signs and Ranks}, Springer.
\url{https://dx.doi.org/10.1007/978-1-4419-0468-3}.\cr
\cr
Nordhausen, K. and Oja, H. (2011), Multivariate L1 Methods:
The Package MNM, \emph{Journal of Statistical Software},
\strong{43}, 1-28. \url{https://doi.org/10.18637/jss.v043.i05}.
}
\seealso{
\code{\link{lambda1s.cv}} for cross-validation of lambda1.
}
